Moved editor components to wrapper component and added routes
